+++ source/tm/move/pp.hpp
@@ -87,7 +87,7 @@ private:
 	}
 
 	// clamped_integer simplifies situations like Pressure and Leppa
-	using current_type = bounded::clamped_integer<0, static_cast<intmax_t>(std::numeric_limits<max_type>::max())>;
+	using current_type = bounded::clamped_integer<0, static_cast<int>(std::numeric_limits<max_type>::max())>;
 	// TODO: Use optional<pair> instead of pair<optional>
 	bounded::optional<max_type> m_max;
 	bounded::optional<current_type> m_current;
